@@ -5,13 +5,13 @@ author: Fayaz
 homepage: https://github.com/fayaz07/ots
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  connectivity: ^3.0.3
+  connectivity_plus: ^1.0.6
 
 dev_dependencies:
   flutter_test:
